feat:bookingDto validation, fix:validation from repositories to services

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -4,7 +4,10 @@ import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.NotFoundException;
+import ru.practicum.shareit.exception.ValidationException;
 import ru.practicum.shareit.item.ItemRepository;
+import ru.practicum.shareit.support.DefaultLocaleMessageSource;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -15,15 +18,24 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements UserService {
     UserRepository userRepository;
     ItemRepository itemRepository;
+    DefaultLocaleMessageSource messageSource;
 
     @Autowired
-    public UserServiceImpl(UserRepository userRepository, ItemRepository itemRepository) {
+    public UserServiceImpl(UserRepository userRepository, ItemRepository itemRepository,
+                           DefaultLocaleMessageSource messageSource) {
         this.userRepository = userRepository;
         this.itemRepository = itemRepository;
+        this.messageSource = messageSource;
     }
 
     @Override
     public UserDto create(UserDto userDto) {
+        if (userDto.getName() == null) {
+            throw new ValidationException("name", messageSource.get("user.UserService.notNullName"));
+        }
+        if (userDto.getEmail() == null) {
+            throw new ValidationException("name", messageSource.get("user.UserService.notNullEmail"));
+        }
         return UserMapper.toUserDto(userRepository.create(UserMapper.toUser(userDto)));
     }
 
@@ -35,17 +47,29 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto readById(Integer id) {
-        return UserMapper.toUserDto(userRepository.readById(id));
+        User user = userRepository.readById(id);
+        if (user == null) {
+            throw new NotFoundException("id", messageSource.get("user.UserService.notFoundById") + ": " + id);
+        }
+        return UserMapper.toUserDto(user);
     }
 
     @Override
     public UserDto update(UserDto userDto) {
-        return UserMapper.toUserDto(userRepository.update(UserMapper.toUser(userDto)));
+        User user = userRepository.update(UserMapper.toUser(userDto));
+        if (user == null) {
+            throw new NotFoundException("id", messageSource.get("user.UserService.notFoundById") + ": "
+                    + userDto.getId());
+        }
+        return UserMapper.toUserDto(user);
     }
 
     @Override
     public void delete(Integer id) {
+        boolean deleted = userRepository.delete(id);
+        if (!deleted) {
+            throw new NotFoundException("id", messageSource.get("user.UserService.notFoundById") + ": " + id);
+        }
         itemRepository.deleteByOwner(id);
-        userRepository.delete(id);
     }
 }
